Extract FeedImageCell test helper to a separate file
