+++ EssentialFeediOSTests/Controllers/FeedViewControllerTests.swift
@@ -347,33 +347,3 @@ class FeedViewControllerTests: XCTestCase {
         }
     }
 }
-
-private extension FeedImageCell {
-    var isShowingLocation: Bool {
-        return !locationContainer.isHidden
-    }
-
-    var isShowingImageLoadingIndicator: Bool {
-        return feedImageContainer.isShimmering
-    }
-
-    var renderedImage: Data? {
-        return feedImageView.image?.pngData()
-    }
-
-    var locationText: String? {
-        return locationLabel.text
-    }
-
-    var descriptionText: String? {
-        return descriptionLabel.text
-    }
-
-    var isShowingRetryAction: Bool {
-        return !feedImageRetryButton.isHidden
-    }
-
-    func simulateRetryAction() {
-        feedImageRetryButton.simulateTap()
-    }
-}
