+++ EssentialFeediOSTests/Helpers/FeedImageCell+TestHelpers.swift
@@ -1,0 +1,39 @@
+//
+//  FeedImageCell+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Viral on 11/05/22.
+//
+
+import UIKit
+import EssentialFeediOS
+
+extension FeedImageCell {
+    var isShowingLocation: Bool {
+        return !locationContainer.isHidden
+    }
+
+    var isShowingImageLoadingIndicator: Bool {
+        return feedImageContainer.isShimmering
+    }
+
+    var renderedImage: Data? {
+        return feedImageView.image?.pngData()
+    }
+
+    var locationText: String? {
+        return locationLabel.text
+    }
+
+    var descriptionText: String? {
+        return descriptionLabel.text
+    }
+
+    var isShowingRetryAction: Bool {
+        return !feedImageRetryButton.isHidden
+    }
+
+    func simulateRetryAction() {
+        feedImageRetryButton.simulateTap()
+    }
+}
